You can now pass in a second, optional, parameter to specify which div to append the #pp_options div. Backwards compatible.
init now accepts an optional css selector for where the #pp_options div should go. The default remains "body", but you can now call it like so:
  $.polypage.init([ 'logged_in', 'admin' ], "#page-controls");

This would result in:
  <body><div id="page-controls"><div id="pp_options"><ul></ul></div></div></body>

If you do not want to specify the defaults, but still specify a selector, call init like so:
  $.polypage.init(null, "#page-controls");

Which will still result in:
  <body><div id="page-controls"><div id="pp_options"><ul></ul></div></div></body>

Not passing in the selector still works, thus this change is backwards compatible.
  $.polypage.init();
  $.polypage.init([ 'logged_in', 'admin' ]);

Both still result in:
  <body><div id="pp_options"><ul></ul></div></body>